find .grok in parent dirs, add `commit` subcommand
`commit` is a new sub-command that generates a git commit message based
on the staged changes in the current git repository. It calls the
`commitMessage` function that invokes `git diff --staged` to get the
diff, checks that it's not too long, and then calls `g.GitCommitMessage`
to generate the commit message.

The code also includes some additional changes to the `answer` function
that make it take a `grokker` object instead of a filename, and some
changes to `chunkStrings` that make it take a string instead of a file.