replace ohcount with Go code to identify programming languages
I need to get rid of the dependency on ohcount by writing or importing
Go code that identifies the programming language of a file.  I've
written getMimeTypes, but the mime type is not enough to determine
the programming language.

Use the go-enry library to replace the ohcount functionality in
getFileLanguages().  See .aidda/response for more information.